+++ tycho-client-py/tests/test_tycho_rpc_client.py
@@ -74,14 +74,13 @@ def test_get_contract_state_returns_expected_result(mock_post, asset_dir):
     mock_post.return_value = mock_response
 
     client = TychoRPCClient()
-    params = ContractStateParams(include_balances=True)
+    params = ContractStateParams()
     result = client.get_contract_state(params)
 
     mock_post.assert_called_once_with(
         "http://0.0.0.0:4242/v1/ethereum/contract_state",
         headers={"accept": "application/json", "Content-Type": "application/json"},
         data="{}",
-        params='{"include_balances": true}',
     )
 
     assert isinstance(result, list)