+++ tycho-indexer/src/extractor/evm/hybrid.rs
@@ -18,8 +18,8 @@ use tycho_core::{
         contract::{Account, AccountDelta},
         protocol::{ComponentBalance, ProtocolComponentState, ProtocolComponentStateDelta},
         token::{CurrencyToken, TokenOwnerStore},
-        Address, Balance, Chain, ChangeType, ExtractionState, ExtractorIdentity, ProtocolType,
-        TxHash,
+        Address, Balance, BlockHash, Chain, ChangeType, ExtractionState, ExtractorIdentity,
+        ProtocolType, TxHash,
     },
     storage::{
         ChainGateway, ContractStateGateway, ExtractionStateGateway, ProtocolGateway, StorageError,
@@ -38,7 +38,7 @@ use crate::{
             protocol_cache::{ProtocolDataCache, ProtocolMemoryCache},
             Block,
         },
-        revert_buffer::RevertBuffer,
+        reorg_buffer::ReorgBuffer,
         BlockUpdateWithCursor, ExtractionError, Extractor, ExtractorMsg,
     },
     pb,
@@ -68,7 +68,7 @@ pub struct HybridContractExtractor<G, T> {
     post_processor: Option<fn(evm::BlockChanges) -> evm::BlockChanges>,
     /// The number of blocks behind the current block to be considered as syncing.
     sync_threshold: u64,
-    revert_buffer: Mutex<RevertBuffer<BlockUpdateWithCursor<evm::BlockChanges>>>,
+    reorg_buffer: Mutex<ReorgBuffer<BlockUpdateWithCursor<evm::BlockChanges>>>,
 }
 
 impl<G, T> HybridContractExtractor<G, T>
@@ -111,7 +111,7 @@ where
                     protocol_types,
                     post_processor,
                     sync_threshold,
-                    revert_buffer: Mutex::new(RevertBuffer::new()),
+                    reorg_buffer: Mutex::new(ReorgBuffer::new()),
                 }
             }
             Ok(cursor) => {
@@ -140,7 +140,7 @@ where
                     protocol_types,
                     post_processor,
                     sync_threshold,
-                    revert_buffer: Mutex::new(RevertBuffer::new()),
+                    reorg_buffer: Mutex::new(ReorgBuffer::new()),
                 }
             }
             Err(err) => return Err(ExtractionError::Setup(err.to_string())),
@@ -237,7 +237,7 @@ where
 
         // Merge stored balances with new ones
         let balances = {
-            let rb = self.revert_buffer.lock().await;
+            let rb = self.reorg_buffer.lock().await;
             let mut balances = self
                 .get_balances(&rb, &balance_request)
                 .await?;
@@ -301,18 +301,18 @@ where
         Ok(())
     }
 
-    /// Returns balances at the tip of the revert buffer.
+    /// Returns balances at the tip of the reorg buffer.
     ///
-    /// Will return the requested balances at the tip of the revert buffer. Might need
+    /// Will return the requested balances at the tip of the reorg buffer. Might need
     /// to go to storage to retrieve balances that are not stored within the buffer.
     async fn get_balances(
         &self,
-        revert_buffer: &RevertBuffer<BlockUpdateWithCursor<evm::BlockChanges>>,
+        reorg_buffer: &ReorgBuffer<BlockUpdateWithCursor<evm::BlockChanges>>,
         reverted_balances_keys: &[(&String, &Bytes)],
     ) -> Result<HashMap<String, HashMap<Bytes, ComponentBalance>>, ExtractionError> {
         // First search in the buffer
         let (buffered_balances, missing_balances_keys) =
-            revert_buffer.lookup_balances(reverted_balances_keys);
+            reorg_buffer.lookup_balances(reverted_balances_keys);
 
         let missing_balances_map: HashMap<String, Vec<Bytes>> = missing_balances_keys
             .into_iter()
@@ -593,14 +593,14 @@ where
         // block finality blockchains.
         let is_syncing = inp.final_block_height >= msg.block.number;
         {
-            // keep revert buffer guard within a limited scope
+            // keep reorg buffer guard within a limited scope
 
-            let mut revert_buffer = self.revert_buffer.lock().await;
-            revert_buffer
+            let mut reorg_buffer = self.reorg_buffer.lock().await;
+            reorg_buffer
                 .insert_block(BlockUpdateWithCursor::new(msg.clone(), inp.cursor.clone()))
                 .map_err(ExtractionError::Storage)?;
 
-            for msg in revert_buffer
+            for msg in reorg_buffer
                 .drain_new_finalized_blocks(inp.final_block_height)
                 .map_err(ExtractionError::Storage)?
             {
@@ -664,12 +664,12 @@ where
             return Ok(None);
         }
 
-        let mut revert_buffer = self.revert_buffer.lock().await;
+        let mut reorg_buffer = self.reorg_buffer.lock().await;
 
         // Purge the buffer
-        let reverted_state = revert_buffer
+        let reverted_state = reorg_buffer
             .purge(block_hash)
-            .map_err(|e| ExtractionError::RevertBufferError(e.to_string()))?;
+            .map_err(|e| ExtractionError::ReorgBufferError(e.to_string()))?;
 
         // Handle created and deleted components
         let (reverted_components_creations, reverted_components_deletions) =
@@ -753,7 +753,7 @@ where
         // Fetch previous values for every reverted states
         // First search in the buffer
         let (buffered_state, missing) =
-            revert_buffer.lookup_account_state(&reverted_account_state_keys_vec);
+            reorg_buffer.lookup_account_state(&reverted_account_state_keys_vec);
 
         // Then for every missing previous values in the buffer, get the data from our db
         let missing_map: HashMap<Bytes, Vec<Bytes>> =
@@ -863,7 +863,7 @@ where
         // Fetch previous values for every reverted states
         // First search in the buffer
         let (buffered_state, missing) =
-            revert_buffer.lookup_protocol_state(&reverted_protocol_state_keys_vec);
+            reorg_buffer.lookup_protocol_state(&reverted_protocol_state_keys_vec);
 
         // Then for every missing previous values in the buffer, get the data from our db
         let missing_map: HashMap<String, Vec<String>> =
@@ -968,13 +968,13 @@ where
         trace!("Reverted balance keys {:?}", &reverted_balances_keys_vec);
 
         let combined_balances = self
-            .get_balances(&revert_buffer, &reverted_balances_keys_vec)
+            .get_balances(&reorg_buffer, &reverted_balances_keys_vec)
             .await?;
 
         let revert_message = BlockAggregatedChanges {
             extractor: self.name.clone(),
             chain: self.chain,
-            block: revert_buffer
+            block: reorg_buffer
                 .get_most_recent_block()
                 .expect("Couldn't find most recent block in buffer during revert"),
             finalized_block_height: reverted_state[0]
@@ -1051,9 +1051,18 @@ impl HybridPgGateway {
     }
 
     #[instrument(skip_all)]
-    async fn save_cursor(&self, new_cursor: &str) -> Result<(), StorageError> {
-        let state =
-            ExtractionState::new(self.name.to_string(), self.chain, None, new_cursor.as_bytes());
+    async fn save_cursor(
+        &self,
+        new_cursor: &str,
+        block_hash: BlockHash,
+    ) -> Result<(), StorageError> {
+        let state = ExtractionState::new(
+            self.name.to_string(),
+            self.chain,
+            None,
+            new_cursor.as_bytes(),
+            block_hash,
+        );
         self.state_gateway
             .save_state(&state)
             .await?;
@@ -1179,7 +1188,8 @@ impl HybridPgGateway {
                 .await?;
         }
 
-        self.save_cursor(new_cursor).await?;
+        self.save_cursor(new_cursor, changes.block.hash.clone())
+            .await?;
 
         let batch_size: usize = if syncing { self.sync_batch_size } else { 0 };
         self.state_gateway
@@ -1764,18 +1774,16 @@ mod test_serial_db {
     use futures03::{stream, StreamExt};
 
     use mockall::mock;
-    use models::contract::TransactionVMUpdates;
     use tycho_core::{
         models::{ContractId, FinancialType, ImplementationType},
         storage::{BlockIdentifier, BlockOrTimestamp},
         traits::TokenOwnerFinding,
     };
-    use tycho_storage::{
-        postgres,
-        postgres::{
-            builder::GatewayBuilder, db_fixtures, db_fixtures::yesterday_midnight,
-            testing::run_against_db,
-        },
+    use tycho_storage::postgres::{
+        self,
+        builder::GatewayBuilder,
+        db_fixtures::{self, yesterday_midnight, yesterday_one_am},
+        testing::run_against_db,
     };
     use web3::types::{H160, H256};
 
@@ -1804,9 +1812,7 @@ mod test_serial_db {
     // VM contract creation fixtures
     const VM_TX_HASH_0: &str = "0x2f6350a292c0fc918afe67cb893744a080dacb507b0cea4cc07437b8aff23cdb";
     const VM_TX_HASH_1: &str = "0x0d9e0da36cf9f305a189965b248fc79c923619801e8ab5ef158d4fd528a291ad";
-    const VM_TX_HASH_2: &str = "0xcf574444be25450fe26d16b85102b241e964a6e01d75dd962203d4888269be3d";
-    const VM_BLOCK_HASH_0: &str =
-        "0x98b4a4fef932b1862be52de218cc32b714a295fae48b775202361a6fa09b66eb";
+
     // Ambient Contract
     const VM_CONTRACT: [u8; 20] = hex_literal::hex!("aaaaaaaaa24eeeb8d57d431224f73832bc34f688");
 
@@ -1914,10 +1920,25 @@ mod test_serial_db {
                 Chain::Ethereum,
                 None,
                 "cursor@420".as_bytes(),
+                Bytes::from_str("88e96d4537bea4d9c05d12549907b32561d3bf31f45aae734cdc119f13406cb6")
+                    .unwrap(),
             );
             evm_gw
                 .start_transaction(&models::blockchain::Block::default(), None)
                 .await;
+            evm_gw
+                .upsert_block(&[models::blockchain::Block {
+                    number: 1,
+                    chain: Chain::Ethereum,
+                    hash: Bytes::from_str(
+                        "88e96d4537bea4d9c05d12549907b32561d3bf31f45aae734cdc119f13406cb6",
+                    )
+                    .unwrap(),
+                    parent_hash: Bytes::default(),
+                    ts: yesterday_one_am(),
+                }])
+                .await
+                .expect("block insertion succeeded");
             evm_gw
                 .save_state(&state)
                 .await
@@ -2128,44 +2149,6 @@ mod test_serial_db {
                     evm::fixtures::transaction02(VM_TX_HASH_1, evm::fixtures::HASH_256_0, 2),
                 ),
             ],
-        }
-    }
-
-    // Allow dead code until reverts are supported again
-    #[allow(dead_code)]
-    fn vm_update02() -> evm::BlockContractChanges {
-        let block = Block::new(
-            1,
-            Chain::Ethereum,
-            VM_BLOCK_HASH_0.parse().unwrap(),
-            H256::zero().into(),
-            "2020-01-01T01:00:00".parse().unwrap(),
-        );
-        evm::BlockContractChanges {
-            extractor: "vm:ambient".to_owned(),
-            chain: Chain::Ethereum,
-            block,
-            finalized_block_height: 0,
-            revert: false,
-            new_tokens: HashMap::new(),
-            tx_updates: vec![TransactionVMUpdates::new(
-                [(
-                    H160(VM_CONTRACT).into(),
-                    AccountDelta::new(
-                        Chain::Ethereum,
-                        VM_CONTRACT.into(),
-                        evm::fixtures::slots([(42, 0xbadbabe)]),
-                        Some(U256::from(2000).into()),
-                        None,
-                        ChangeType::Update,
-                    ),
-                )]
-                .into_iter()
-                .collect(),
-                HashMap::new(),
-                HashMap::new(),
-                evm::fixtures::transaction02(VM_TX_HASH_2, VM_BLOCK_HASH_0, 1),
-            )],
         }
     }