+++ tycho-storage/migrations/2024-09-06_v0.16.2/up.sql
@@ -1,0 +1,3 @@
+DROP TABLE IF EXISTS component_balance_old;
+DROP TABLE IF EXISTS contract_storage_old;
+DROP TABLE IF EXISTS protocol_state_old;