+++ tycho-storage/src/postgres/mod.rs
@@ -603,8 +603,8 @@ pub mod testing {
         let tables = vec![
             // put block early so most FKs cascade, it would
             // be better to find the correct order tough.
+            "extraction_state",
             "block",
-            "protocol_calls_contract",
             "contract_storage",
             "contract_code",
             "account_balance",
@@ -615,7 +615,6 @@ pub mod testing {
             "account",
             "protocol_state",
             "protocol_component",
-            "extraction_state",
             "protocol_type",
             "protocol_system",
             "transaction",
@@ -625,7 +624,7 @@ pub mod testing {
             sql_query(format!("DELETE FROM {};", t))
                 .execute(conn)
                 .await
-                .unwrap_or_else(|_| panic!("Error truncating {} table", t));
+                .unwrap_or_else(|e| panic!("Error truncating {} table: {}", t, e));
         }
         dbg!("Teardown completed");
     }