+++ tycho-storage/src/postgres/schema.rs
@@ -64,6 +64,34 @@ diesel::table! {
 }
 
 diesel::table! {
+    component_balance (token_id, protocol_component_id, valid_to) {
+        token_id -> Int8,
+        new_balance -> Bytea,
+        previous_value -> Bytea,
+        balance_float -> Float8,
+        modify_tx -> Int8,
+        protocol_component_id -> Int8,
+        inserted_ts -> Timestamptz,
+        valid_from -> Timestamptz,
+        valid_to -> Timestamptz,
+    }
+}
+
+diesel::table! {
+    component_balance_default (token_id, protocol_component_id, valid_to) {
+        token_id -> Int8,
+        new_balance -> Bytea,
+        previous_value -> Bytea,
+        balance_float -> Float8,
+        modify_tx -> Int8,
+        protocol_component_id -> Int8,
+        inserted_ts -> Timestamptz,
+        valid_from -> Timestamptz,
+        valid_to -> Timestamptz,
+    }
+}
+
+diesel::table! {
     component_tvl (id) {
         id -> Int8,
         protocol_component_id -> Int8,
@@ -88,6 +116,36 @@ diesel::table! {
 }
 
 diesel::table! {
+    contract_storage (account_id, slot, valid_to) {
+        slot -> Bytea,
+        value -> Nullable<Bytea>,
+        previous_value -> Nullable<Bytea>,
+        account_id -> Int8,
+        modify_tx -> Int8,
+        ordinal -> Int8,
+        valid_from -> Timestamptz,
+        valid_to -> Timestamptz,
+        inserted_ts -> Timestamptz,
+        modified_ts -> Timestamptz,
+    }
+}
+
+diesel::table! {
+    contract_storage_default (account_id, slot, valid_to) {
+        slot -> Bytea,
+        value -> Nullable<Bytea>,
+        previous_value -> Nullable<Bytea>,
+        account_id -> Int8,
+        modify_tx -> Int8,
+        ordinal -> Int8,
+        valid_from -> Timestamptz,
+        valid_to -> Timestamptz,
+        inserted_ts -> Timestamptz,
+        modified_ts -> Timestamptz,
+    }
+}
+
+diesel::table! {
     extraction_state (id) {
         id -> Int8,
         #[max_length = 255]
@@ -99,18 +157,7 @@ diesel::table! {
         attributes -> Nullable<Jsonb>,
         inserted_ts -> Timestamptz,
         modified_ts -> Timestamptz,
-    }
-}
-
-diesel::table! {
-    protocol_calls_contract (id) {
-        id -> Int8,
-        protocol_component_id -> Int8,
-        account_id -> Int8,
-        valid_from -> Timestamptz,
-        valid_to -> Nullable<Timestamptz>,
-        inserted_ts -> Timestamptz,
-        modified_ts -> Timestamptz,
+        block_id -> Int8,
     }
 }
 
@@ -147,6 +194,34 @@ diesel::table! {
         token_id -> Int8,
         inserted_ts -> Timestamptz,
         modified_ts -> Timestamptz,
+    }
+}
+
+diesel::table! {
+    protocol_state (protocol_component_id, attribute_name, valid_to) {
+        attribute_name -> Varchar,
+        attribute_value -> Bytea,
+        previous_value -> Nullable<Bytea>,
+        modify_tx -> Int8,
+        valid_from -> Timestamptz,
+        valid_to -> Timestamptz,
+        inserted_ts -> Timestamptz,
+        modified_ts -> Timestamptz,
+        protocol_component_id -> Int8,
+    }
+}
+
+diesel::table! {
+    protocol_state_default (protocol_component_id, attribute_name, valid_to) {
+        attribute_name -> Varchar,
+        attribute_value -> Bytea,
+        previous_value -> Nullable<Bytea>,
+        modify_tx -> Int8,
+        valid_from -> Timestamptz,
+        valid_to -> Timestamptz,
+        inserted_ts -> Timestamptz,
+        modified_ts -> Timestamptz,
+        protocol_component_id -> Int8,
     }
 }
 
@@ -219,19 +294,32 @@ diesel::joinable!(account -> chain (chain_id));
 diesel::joinable!(account_balance -> account (account_id));
 diesel::joinable!(account_balance -> transaction (modify_tx));
 diesel::joinable!(block -> chain (chain_id));
+diesel::joinable!(component_balance -> protocol_component (protocol_component_id));
+diesel::joinable!(component_balance -> token (token_id));
+diesel::joinable!(component_balance -> transaction (modify_tx));
+diesel::joinable!(component_balance_default -> protocol_component (protocol_component_id));
+diesel::joinable!(component_balance_default -> token (token_id));
+diesel::joinable!(component_balance_default -> transaction (modify_tx));
 diesel::joinable!(component_tvl -> protocol_component (protocol_component_id));
 diesel::joinable!(contract_code -> account (account_id));
 diesel::joinable!(contract_code -> transaction (modify_tx));
+diesel::joinable!(contract_storage -> account (account_id));
+diesel::joinable!(contract_storage -> transaction (modify_tx));
+diesel::joinable!(contract_storage_default -> account (account_id));
+diesel::joinable!(contract_storage_default -> transaction (modify_tx));
+diesel::joinable!(extraction_state -> block (block_id));
 diesel::joinable!(extraction_state -> chain (chain_id));
-diesel::joinable!(protocol_calls_contract -> account (account_id));
-diesel::joinable!(protocol_calls_contract -> protocol_component (protocol_component_id));
 diesel::joinable!(protocol_component -> chain (chain_id));
 diesel::joinable!(protocol_component -> protocol_system (protocol_system_id));
 diesel::joinable!(protocol_component -> protocol_type (protocol_type_id));
 diesel::joinable!(protocol_component_holds_contract -> contract_code (contract_code_id));
 diesel::joinable!(protocol_component_holds_contract -> protocol_component (protocol_component_id));
-diesel::joinable!(protocol_component_holds_token -> protocol_component (protocol_component_id));
 diesel::joinable!(protocol_component_holds_token -> token (token_id));
+diesel::joinable!(protocol_component_holds_token -> protocol_component (protocol_component_id));
+diesel::joinable!(protocol_state -> protocol_component (protocol_component_id));
+diesel::joinable!(protocol_state -> transaction (modify_tx));
+diesel::joinable!(protocol_state_default -> protocol_component (protocol_component_id));
+diesel::joinable!(protocol_state_default -> transaction (modify_tx));
 diesel::joinable!(token -> account (account_id));
 diesel::joinable!(token_price -> token (token_id));
 diesel::joinable!(transaction -> block (block_id));
@@ -241,123 +329,21 @@ diesel::allow_tables_to_appear_in_same_query!(
     account_balance,
     block,
     chain,
+    component_balance,
+    component_balance_default,
     component_tvl,
     contract_code,
+    contract_storage,
+    contract_storage_default,
     extraction_state,
-    protocol_calls_contract,
     protocol_component,
     protocol_component_holds_contract,
     protocol_component_holds_token,
+    protocol_state,
+    protocol_state_default,
     protocol_system,
     protocol_type,
     token,
     token_price,
     transaction,
-    component_balance,
-    component_balance_default,
-    contract_storage,
-    contract_storage_default,
-    protocol_state,
-    protocol_state_default
 );
-
-diesel::table! {
-    component_balance (protocol_component_id, token_id, modify_tx) {
-        token_id -> Int8,
-        new_balance -> Bytea,
-        previous_value -> Bytea,
-        balance_float -> Float8,
-        modify_tx -> Int8,
-        protocol_component_id -> Int8,
-        inserted_ts -> Timestamptz,
-        valid_from -> Timestamptz,
-        valid_to -> Timestamptz,
-    }
-}
-
-diesel::table! {
-    component_balance_default (protocol_component_id, token_id, modify_tx) {
-        token_id -> Int8,
-        new_balance -> Bytea,
-        previous_value -> Bytea,
-        balance_float -> Float8,
-        modify_tx -> Int8,
-        protocol_component_id -> Int8,
-        inserted_ts -> Timestamptz,
-        valid_from -> Timestamptz,
-        valid_to -> Timestamptz,
-    }
-}
-
-diesel::table! {
-    contract_storage (account_id, slot, modify_tx) {
-        slot -> Bytea,
-        value -> Nullable<Bytea>,
-        previous_value -> Nullable<Bytea>,
-        account_id -> Int8,
-        modify_tx -> Int8,
-        ordinal -> Int8,
-        valid_from -> Timestamptz,
-        valid_to -> Timestamptz,
-        inserted_ts -> Timestamptz,
-        modified_ts -> Timestamptz,
-    }
-}
-
-diesel::table! {
-    contract_storage_default (account_id, slot, modify_tx) {
-        slot -> Bytea,
-        value -> Nullable<Bytea>,
-        previous_value -> Nullable<Bytea>,
-        account_id -> Int8,
-        modify_tx -> Int8,
-        ordinal -> Int8,
-        valid_from -> Timestamptz,
-        valid_to -> Timestamptz,
-        inserted_ts -> Timestamptz,
-        modified_ts -> Timestamptz,
-    }
-}
-
-diesel::table! {
-    protocol_state (protocol_component_id, attribute_name, modify_tx) {
-        modify_tx -> Int8,
-        valid_from -> Timestamptz,
-        valid_to -> Timestamptz,
-        inserted_ts -> Timestamptz,
-        modified_ts -> Timestamptz,
-        protocol_component_id -> Int8,
-        attribute_name -> Varchar,
-        attribute_value -> Bytea,
-        previous_value -> Nullable<Bytea>,
-    }
-}
-
-diesel::table! {
-    protocol_state_default (protocol_component_id, attribute_name, modify_tx) {
-        modify_tx -> Int8,
-        valid_from -> Timestamptz,
-        valid_to -> Timestamptz,
-        inserted_ts -> Timestamptz,
-        modified_ts -> Timestamptz,
-        protocol_component_id -> Int8,
-        attribute_name -> Varchar,
-        attribute_value -> Bytea,
-        previous_value -> Nullable<Bytea>,
-    }
-}
-
-diesel::joinable!(component_balance -> protocol_component (protocol_component_id));
-diesel::joinable!(component_balance -> token (token_id));
-diesel::joinable!(component_balance -> transaction (modify_tx));
-diesel::joinable!(component_balance_default -> protocol_component (protocol_component_id));
-diesel::joinable!(component_balance_default -> token (token_id));
-diesel::joinable!(component_balance_default -> transaction (modify_tx));
-diesel::joinable!(contract_storage -> account (account_id));
-diesel::joinable!(contract_storage -> transaction (modify_tx));
-diesel::joinable!(contract_storage_default -> account (account_id));
-diesel::joinable!(contract_storage_default -> transaction (modify_tx));
-diesel::joinable!(protocol_state -> protocol_component (protocol_component_id));
-diesel::joinable!(protocol_state -> transaction (modify_tx));
-diesel::joinable!(protocol_state_default -> protocol_component (protocol_component_id));
-diesel::joinable!(protocol_state_default -> transaction (modify_tx));